+++ ContentPatcher/Framework/Migrations/AggregateMigration.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using ContentPatcher.Framework.Conditions;
 using ContentPatcher.Framework.ConfigModels;
 using ContentPatcher.Framework.Lexing.LexTokens;
 using ContentPatcher.Framework.Tokens.Json;
@@ -40,10 +41,7 @@ namespace ContentPatcher.Framework.Migrations
             this.Migrations = migrations.Where(m => m.Version.IsNewerThan(version)).ToArray();
         }
 
-        /// <summary>Migrate a content pack.</summary>
-        /// <param name="content">The content pack data to migrate.</param>
-        /// <param name="error">An error message which indicates why migration failed.</param>
-        /// <returns>Returns whether the content pack was successfully migrated.</returns>
+        /// <inheritdoc />
         public bool TryMigrate(ContentConfig content, out string error)
         {
             // validate format version
@@ -65,10 +63,7 @@ namespace ContentPatcher.Framework.Migrations
             return true;
         }
 
-        /// <summary>Migrate a lexical token.</summary>
-        /// <param name="lexToken">The lexical token to migrate.</param>
-        /// <param name="error">An error message which indicates why migration failed (if any).</param>
-        /// <returns>Returns whether migration succeeded.</returns>
+        /// <inheritdoc />
         public bool TryMigrate(ILexToken lexToken, out string error)
         {
             // apply migrations
@@ -83,10 +78,7 @@ namespace ContentPatcher.Framework.Migrations
             return true;
         }
 
-        /// <summary>Migrate a tokenized string.</summary>
-        /// <param name="tokenStr">The tokenized string to migrate.</param>
-        /// <param name="error">An error message which indicates why migration failed (if any).</param>
-        /// <returns>Returns whether migration succeeded.</returns>
+        /// <inheritdoc />
         public bool TryMigrate(IManagedTokenString tokenStr, out string error)
         {
             // apply migrations
@@ -101,16 +93,28 @@ namespace ContentPatcher.Framework.Migrations
             return true;
         }
 
-        /// <summary>Migrate a tokenized JSON structure.</summary>
-        /// <param name="tokenStructure">The tokenized JSON structure to migrate.</param>
-        /// <param name="error">An error message which indicates why migration failed (if any).</param>
-        /// <returns>Returns whether migration succeeded.</returns>
+        /// <inheritdoc />
         public bool TryMigrate(TokenizableJToken tokenStructure, out string error)
         {
             // apply migrations
             foreach (IMigration migration in this.Migrations)
             {
                 if (!migration.TryMigrate(tokenStructure, out error))
+                    return false;
+            }
+
+            // no issues found
+            error = null;
+            return true;
+        }
+
+        /// <inheritdoc />
+        public bool TryMigrate(Condition condition, out string error)
+        {
+            // apply migrations
+            foreach (IMigration migration in this.Migrations)
+            {
+                if (!migration.TryMigrate(condition, out error))
                     return false;
             }