+++ ContentPatcher/Framework/Migrations/Migration_1_9.cs
@@ -15,10 +15,7 @@ namespace ContentPatcher.Framework.Migrations
         public Migration_1_9()
             : base(new SemanticVersion(1, 9, 0)) { }
 
-        /// <summary>Migrate a lexical token.</summary>
-        /// <param name="lexToken">The lexical token to migrate.</param>
-        /// <param name="error">An error message which indicates why migration failed (if any).</param>
-        /// <returns>Returns whether migration succeeded.</returns>
+        /// <inheritdoc />
         public override bool TryMigrate(ILexToken lexToken, out string error)
         {
             if (!base.TryMigrate(lexToken, out error))