+++ TractorMod/Framework/BaseAttachment.cs
@@ -327,5 +327,36 @@ namespace Pathoschild.Stardew.TractorMod.Framework
             isCoveredByObj = false;
             return false;
         }
+
+        /// <summary>Break open a container using a tool, if applicable.</summary>
+        /// <param name="tile">The tile position</param>
+        /// <param name="tileObj">The object on the tile.</param>
+        /// <param name="tool">The tool selected by the player (if any).</param>
+        /// <param name="location">The current location.</param>
+        protected bool TryBreakContainer(Vector2 tile, SObject tileObj, Tool tool, GameLocation location)
+        {
+            if (tileObj is BreakableContainer)
+                return tileObj.performToolAction(tool, location);
+
+            return false;
+        }
+
+        /// <summary>Cancel the current player animation if it matches one of the given IDs.</summary>
+        /// <param name="player">The player to change.</param>
+        /// <param name="animationIds">The animation IDs to detect.</param>
+        protected void CancelAnimation(Farmer player, params int[] animationIds)
+        {
+            int animationId = this.Reflection.GetField<int>(player.FarmerSprite, "currentSingleAnimation").GetValue();
+            foreach (int id in animationIds)
+            {
+                if (id == animationId)
+                {
+                    player.completelyStopAnimatingOrDoingAction();
+                    player.forceCanMove();
+
+                    break;
+                }
+            }
+        }
     }
 }