+++ TractorMod/release-notes.md
@@ -1,4 +1,11 @@
 # Release notes
+## 4.12
+Released 05 December 2020.
+
+* The pickaxe now collects spawned mine items like quartz.
+* The scythe now harvests palm tree coconuts and tree seeds.
+* Fixed scythe harvesting spring onions based on 'harvest crops' option instead of 'harvest forage'.
+
 ## 4.11.3
 Released 15 October 2020.